feat: Vehicles list (Povillas' version)
