+++ src/hooks/useVehicle.jsx
@@ -1,16 +1,9 @@
 import { useState, useEffect } from "react";
-import { useNavigate, useLocation } from 'react-router-dom';
 
 function useVehicle() {
-    const location = useLocation();
     const [errors, setErrors] = useState({})
     const [loading, setLoading] = useState(false);
     const [data, setData] = useState([])
-    const navigate = useNavigate()
-
-    useEffect(() => {
-      if (location.pathname === '/vehicles' && data.length === 0) getVehicles();
-    }, [location, data])
 
     function createVehicle(data) {
         setLoading(true)
@@ -26,21 +19,9 @@ function useVehicle() {
           .finally(() => setLoading(false))
     }
 
-    function getVehicles() {
-      setLoading(true);
-
-      return axios.get('vehicles')
-        .then(response => {
-          setData(response.data.data)
-        })
-        .catch((err) => console.error(err))
-        .finally(() => setLoading(false));
-    }
-
     return {
         vehicle: { data, setData, errors, loading },
         createVehicle,
-        getVehicles,
     }
 }
